fix 3 наверное

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,13 +1,20 @@
 package ru.yandex.practicum.filmorate.service;
 
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.FilmNotFoundException;
+import ru.yandex.practicum.filmorate.exception.FilmReleaseException;
 
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
 import java.util.List;
 
+
+@Slf4j
 @Service
 public class FilmService {
     private final FilmStorage filmStorage;
@@ -25,10 +32,16 @@ public class FilmService {
     }
 
     public Film update(Film film) throws FilmNotFoundException {
+        if (filmStorage.getFilm(film.getId()) == null) {
+            throw new FilmNotFoundException("Film with this ID doesn't exist.");
+        }
         return filmStorage.update(film);
     }
 
     public Film getFilm(int id) {
+        if (filmStorage.getFilm(id) == null) {
+            throw new FilmNotFoundException("Film doesn't exist.");
+        }
         return filmStorage.getFilm(id);
     }
 
@@ -37,10 +50,11 @@ public class FilmService {
     }
 
     public Film deleteLike(int id, int userId) {
+        if (filmStorage.getFilm(id) == null) {
+            throw new FilmNotFoundException("Film doesn't exist.");
+        }
         return filmStorage.deleteLike(id, userId);
     }
 
-    public List<Film> findPopularFilms(int count) {
-        return filmStorage.findPopularFilms(count);
-    }
+
 }
